D8 Flow Directions

D8 Flow Directions

Usage

taudem_d8flowdir(
  input_elevation_grid,
  output_d8flowdir_grid = NULL,
  output_d8slopes_grid = NULL,
  n_processes = getOption("traudem.n_processes", 1),
  quiet = getOption("traudem.quiet", FALSE)
)

Arguments

input_elevation_grid

Pit filled elevation input data

output_d8flowdir_grid

D8 flow directions output

output_d8slopes_grid

D8 slopes output

n_processes

Number of processes for mpiexec. If NULL TauDEM is called without mpiexec.

quiet

If FALSE output from TauDEM CLI is suppressed.

Value

List with the two output filenames

Details

See https://hydrology.usu.edu/taudem/taudem5/help53/D8FlowDirections.html

Examples

# \dontrun{
test_dir <- withr::local_tempdir()
dir.create(test_dir)
 file.copy(
   system.file("test-data", "MED_01_01.tif", package = "traudem"),
   file.path(test_dir, "MED_01_01.tif")
 )
#> [1] TRUE
filled_pit <- taudem_pitremove(file.path(test_dir, "MED_01_01.tif"))
#> PitRemove version 5.3.9
#> Input file /tmp/Rtmpfcdi01/file3eea77a3e9b8/MED_01_01.tif has projected coordinate system.
#> Nodata value input to create partition from file: -340282299999999994960115009090224128000.000000
#> Nodata value recast to float used in partition raster: -340282306073709652508363335590014353408.000000
#> Processes: 1
#> Header read time: 0.003650
#> Data read time: 0.003797
#> Compute time: 0.050725
#> Write time: 0.014980
#> Total time: 0.073152
#> This run may take on the order of 1 minutes to complete.
#> This estimate is very approximate. 
#> Run time is highly uncertain as it depends on the complexity of the input data 
#> and speed and memory of the computer. This estimate is based on our testing on 
#> a dual quad core Dell Xeon E5405 2.0GHz PC with 16GB RAM.
outputs <- taudem_d8flowdir(filled_pit)
#> D8FlowDir version 5.3.9
#> Input file /tmp/Rtmpfcdi01/file3eea77a3e9b8/MED_01_01fel.tif has projected coordinate system.
#> Nodata value input to create partition from file: -300000000549775575777803994281145270272.000000
#> Nodata value recast to float used in partition raster: -300000000549775575777803994281145270272.000000
#> Processors: 1
#> Header read time: 0.003341
#> Data read time: 0.007174
#> Compute Slope time: 0.050894
#> Write Slope time: 0.009835
#> Resolve Flat time: 5.614253
#> Write Flat time: 0.008085
#> Total time: 5.693581
#> This run may take on the order of 1 minutes to complete.
#> This estimate is very approximate. 
#> Run time is highly uncertain as it depends on the complexity of the input data 
#> and speed and memory of the computer. This estimate is based on our testing on 
#> a dual quad core Dell Xeon E5405 2.0GHz PC with 16GB RAM.
#> All slopes evaluated. 151718 flats to resolve.
#> Draining flats towards lower adjacent terrain
#> ......................................................................................................................................................................................................................
#> Draining flats away from higher adjacent terrain
#> ........................................
#> Setting directions
#> Iteration complete. Number of flats remaining: 283
#> Draining flats towards lower adjacent terrain
#> .......
#> Draining flats away from higher adjacent terrain
#> .....
#> Setting directions
outputs
#> $output_d8flowdir_grid
#> [1] "/tmp/Rtmpfcdi01/file3eea77a3e9b8/MED_01_01feld8flowdir.tif"
#> 
#> $output_d8slopes_grid
#> [1] "/tmp/Rtmpfcdi01/file3eea77a3e9b8/MED_01_01feld8slopes.tif"
#> 
# }