TauDEM situation report

Checks installation of TauDEM and provides useful hints.

Usage

taudem_sitrep()

can_register_taudem()

Value

For taudem_sitrep(): None.

For can_register_taudem(): A logical scalar.

TauDEM installation and registration

Once you have installed TauDEM, add an environment variable pointing to the correct path. For instance on Ubuntu it could be adding this line in .Renviron (see usethis::edit_r_environ()) and then re-starting R:

TAUDEM_PATH='/usr/local/taudem'

or, for just the session, running this line of R code:

Sys.setenv(TAUDEM_PATH = "/usr/local/taudem")

Examples

if (FALSE) { # interactive() && traudem::can_register_taudem()
try(taudem_sitrep(), silent = TRUE)
}
can_register_taudem()
#> [1] TRUE