TauDEM situation report

Checks installation of TauDEM and provides useful hints.

Usage

taudem_sitrep()

Value

None

TauDEM installation and registration

Once you have installed TauDEM add an environment variable pointing to the correct path. For instance on Ubuntu it could be adding this line in .Renviron (see usethis::edit_r_environ()) and then re-starting R:

TAUDEM_PATH='/usr/local/taudem'

or, for just the session, running this line of R code:

Sys.setenv(TAUDEM_PATH='/usr/local/taudem')

Examples

# \dontrun{
try(taudem_sitrep(), silent = TRUE)
#> ✔ Found GDAL version 3.0.4.
#> ✔ Found HYDRA build details: (MPI).
#> ✔ Found TauDEM path (/home/runner/TauDEM/bin).
#> ✔ Found TauDEM executables directory (/home/runner/TauDEM/bin).
#> ✔ Found all TauDEM executables.
#> ℹ Testing TauDEM on an example file...
#> ── TauDEM output ───────────────────────────────────────────────────────────────
#> PitRemove version 5.3.9
#> Input file MED_01_01.tif has projected coordinate system.
#> Nodata value input to create partition from file: -340282299999999994960115009090224128000.000000
#> Nodata value recast to float used in partition raster: -340282306073709652508363335590014353408.000000
#> Processes: 1
#> Header read time: 0.003391
#> Data read time: 0.003951
#> Compute time: 0.064464
#> Write time: 0.015836
#> Total time: 0.087642
#> This run may take on the order of 1 minutes to complete.
#> This estimate is very approximate. 
#> Run time is highly uncertain as it depends on the complexity of the input data 
#> and speed and memory of the computer. This estimate is based on our testing on 
#> a dual quad core Dell Xeon E5405 2.0GHz PC with 16GB RAM.
#> ── End of TauDEM output ────────────────────────────────────────────────────────
#> ✔ Was able to launch a TauDEM example!
#> ! Make sure you see no serious error message above.
# }